refactor: database to use bigint instead of uuid and numeric

+++ server/utils/db/postgres/schema/tables/reactionRoles.ts
@@ -1,4 +1,4 @@
-import { colJson, numeric, primaryKey, varchar } from '@kmods/drizzle-pg/pg-core';
+import { bigint, colJson, primaryKey, varchar } from '@kmods/drizzle-pg/pg-core';
 import { kbot2Schema } from '../pgSchema';
 import { defaultGuildFields } from './guilds';
 
@@ -8,16 +8,16 @@ export type ReactionRole = {
 };
 
 export const scReactionRoles = kbot2Schema.table('discord_guild_reaction_roles_message', {
-	message_id: numeric('message_id').notNull().primaryKey(),
+	message_id: bigint('message_id').notNull().primaryKey(),
 	...defaultGuildFields,
-	channel_id: numeric('channel_id').notNull(),
+	channel_id: bigint('channel_id').notNull(),
 	name: varchar('name', { length: 128 }).notNull().default('')
 });
 
 export const scReactionRolesEmojies = kbot2Schema.table(
 	'discord_guild_reaction_roles_message_emojies',
 	{
-		message_id: numeric('message_id').references(
+		message_id: bigint('message_id').references(
 			() => {
 				return scReactionRoles.message_id;
 			},
